Revert "Revert "Car, ParkingPermit, ParkingTransaction, ParkingOffice refactor.""
This reverts commit 2e9a1a21042077016114fc64f37a3537372a0090.

+++ src/test/java/com/mycompany/parkingsystem/CarTypeTest.java
@@ -24,12 +24,12 @@ public class CarTypeTest {
     public void testValuesSUV() {
         try {
             System.out.println("values");
-            long value = 1000;
+            double value = 1000;
             CarType inst = null;
             inst = inst.valueOf("SUV");
             inst.setValue(value);
-            long expresult = 1000;
-            long result = inst.showValue(inst);
+            double expresult = 1000;
+            double result = inst.showValue(inst);
             assertEquals(expresult,result);
         } catch (Exception e) {
             fail("SUV returned wrong value");
@@ -44,12 +44,12 @@ public class CarTypeTest {
     public void testValuesCOMPACT() {
         try {
             System.out.println("values");
-            long value = 1000;
+            double value = 1000;
             CarType inst = null;
             inst = inst.valueOf("COMPACT");
             inst.setValue(value);
-            long expresult = 800;
-            long result = inst.showValue(inst);
+            double expresult = 800;
+            double result = inst.showValue(inst);
             assertEquals(expresult,result);
         } catch (Exception e) {
             fail("COMPACT returned wrong value");
