+++ src/test/java/com/mycompany/parkingsystem/ParkingLotTest.java
@@ -24,13 +24,13 @@ public class ParkingLotTest {
     public void testGetDailyRateCOMPACT() {
         try {
             System.out.println("getDailyRate");
-            long amt = 10000;
+            double amt = 10000;
             CarType ct = CarType.COMPACT;
             ParkingLot instance = new ParkingLot();
             instance.setMaxCost(amt);
             Money m = instance.getDailyRate(ct);
-            long result = m.getAmount();
-            long expresult = 8000;
+            double result = m.getAmount();
+            double expresult = 8000;
             assertEquals(expresult, result);
         } catch (Exception e) {
             fail("getDailyRate returned unexpected results COMPACT");
@@ -44,13 +44,13 @@ public class ParkingLotTest {
     public void testGetDailyRateSUV() {
         try {
             System.out.println("getDailyRate");
-            long amt = 1000;
+            double amt = 1000;
             CarType ct = CarType.SUV;
             ParkingLot instance = new ParkingLot();
             instance.setMaxCost(amt);
             Money m = instance.getDailyRate(ct);
-            long result = m.getAmount();
-            long expresult = amt;
+            double result = m.getAmount();
+            double expresult = amt;
             assertEquals(expresult, result);
         } catch (Exception e) {
             fail("getDailyRate returned unexpected results SUV");
